ANDROID: db845c: disable VIDEO_QCOM_VENUS
Upstream commit 687bfbba5a1c ("media: venus: Add support for static
video encoder/decoder declarations") introduced a new feature in the
qcom venus driver that depends on the OF_DYNAMIC Kconfig option.

This dependency causes build errors for db845c as OF_DYNAMIC is not
currently enabled on the gki_defconfig. Furthermore, enabling this
option is complicated as it's a bool type that can only be selected
through other configs.

Lets disable the venus driver temporarily while a proper solution for
this dependency is investigated.

Bug: 408097272
Change-Id: I92fac43c6cf28358421b4258f1d9f03a1138c368
Signed-off-by: Carlos Llamas <cmllamas@google.com>